Changed procsci to be enum, updated migration logic to match

+++ dbmigration/versions/010_change_procsci_to_enum.py
@@ -1,19 +1,33 @@
 from sqlalchemy import *
 from migrate import *
 
+from fits_storage.orm.header import PROCSCI_ENUM
+
+
 def upgrade(migrate_engine):
     meta = MetaData(bind = migrate_engine)
     header = Table('header', meta, autoload=True)
 
-    procsci    = Column('procsci', String(4))
+    PROCSCI_ENUM.create(migrate_engine)
+    if hasattr(header.c, 'procsci'):
+        header.c.procsci.drop()
+    procsci = Column('procsci', PROCSCI_ENUM)
 
     procsci.create(header)
+
 
 def downgrade(migrate_engine):
     meta = MetaData(bind = migrate_engine)
     header = Table('header', meta, autoload=True)
 
-    header.c.procsci.drop()
+    if hasattr(header.c, 'procsci'):
+        header.c.procsci.drop()
+    PROCSCI_ENUM.drop(migrate_engine)
+
+    # recreate varchar col
+    procsci    = Column('procsci', String(4))
+
+    procsci.create(header)
 
 
 # putting this here since most new migrations begin as a copy/paste
